Optimize shell script
On busy servers, running `find /proc` with no flags can not only take a
relatively long time, but may cause `proc_inode_cache` in slab to
balloon, which in turn may force the kernel to go on an OOM killing
spree.

Since this is only required to check that chkrootkit is running on a
live system, limit the `find` run to only the `/proc`'s immediate
children.